+++ src/DataTable/DataTable.js
@@ -1,10 +1,14 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
-import s from './DataTable.scss';
+import oldStyle from './DataTable.old.scss';
+import newStyle from './DataTable.scss';
+import typography from '../Typography/Typography.scss';
 import classNames from 'classnames';
 import InfiniteScroll from './InfiniteScroll';
 import WixComponent from '../BaseComponents/WixComponent';
 import ArrowVertical from '../Icons/dist/components/ArrowVertical';
+import SortByArrowDown from 'wix-ui-icons-common/system/SortByArrowDown';
+import SortByArrowUp from 'wix-ui-icons-common/system/SortByArrowUp';
 import {Animator} from 'wix-animations';
 import InfoCircle from 'wix-ui-icons-common/InfoCircle';
 import Tooltip from '../Tooltip/Tooltip';
@@ -12,7 +16,7 @@ import omit from 'lodash/omit';
 
 export const DataTableHeader = props => (
   <div>
-    <table style={{width: props.width}} className={s.table}>
+    <table style={{width: props.width}} className={newStyle.table}>
       <TableHeader {...props}/>
     </table>
   </div>
@@ -23,6 +27,7 @@ DataTableHeader.propTypes = {
 };
 
 class DataTable extends WixComponent {
+
   constructor(props) {
     super(props);
     let state = {selectedRows: {}};
@@ -30,6 +35,10 @@ class DataTable extends WixComponent {
       state = {...state, ...this.createInitialScrollingState(props)};
     }
     this.state = state;
+  }
+
+  get style() {
+    return this.props.newDesign ? newStyle : oldStyle;
   }
 
   componentWillReceiveProps(nextProps) {
@@ -95,7 +104,7 @@ class DataTable extends WixComponent {
     const style = {width: this.props.width};
     return (
       <div>
-        <table id={this.props.id} style={style} className={s.table}>
+        <table id={this.props.id} style={style} className={this.style.table}>
           {!this.props.hideHeader &&
           <TableHeader {...this.props}/>}
           {this.renderBody(rowsToRender)}
@@ -138,11 +147,11 @@ class DataTable extends WixComponent {
     });
 
     if (onRowClick) {
-      rowClasses.push(s.clickableDataRow);
+      rowClasses.push(this.style.clickableDataRow);
     }
 
     if (rowDetails) {
-      rowClasses.push(s.animatedDataRow);
+      rowClasses.push(this.style.animatedDataRow);
     }
 
     if (rowDataHook) {
@@ -169,13 +178,13 @@ class DataTable extends WixComponent {
       const showDetails = !!this.state.selectedRows[rowNum];
 
       rowsToRender.push(
-        <tr key={`${rowNum}_details`} className={classNames(s.rowDetails)}>
+        <tr key={`${rowNum}_details`} className={classNames(this.style.rowDetails)}>
           <td
             data-hook={`${rowNum}_details`}
-            className={classNames(s.details, showDetails ? s.active : '')}
+            className={classNames(this.style.details, showDetails ? this.style.active : '')}
             colSpan={this.props.columns.length}
             >
-            <div className={classNames(s.rowDetailsInner)}>
+            <div className={classNames(this.style.rowDetailsInner)}>
               <Animator show={showDetails} height>
                 {rowDetails(rowData, rowNum)}
               </Animator>
@@ -189,7 +198,11 @@ class DataTable extends WixComponent {
   };
 
   renderCell = (rowData, column, rowNum, colNum) => {
-    const classes = classNames({[s.important]: column.important});
+    const classes = classNames(
+      {[typography.t1]: this.props.newDesign},
+      {[this.style.important]: column.important},
+      {[this.style.largeVerticalPadding]: this.props.rowVerticalPadding === 'large'},
+      {[this.style.mediumVerticalPadding]: this.props.rowVerticalPadding !== 'large'});
     const width = rowNum === 0 && this.props.hideHeader ? column.width : undefined;
 
     return (<td
@@ -222,6 +235,7 @@ class DataTable extends WixComponent {
 }
 
 class TableHeader extends Component {
+
   static propTypes = {
     onSortClick: PropTypes.func,
     thPadding: PropTypes.string,
@@ -231,14 +245,30 @@ class TableHeader extends Component {
     thColor: PropTypes.string,
     thOpacity: PropTypes.string,
     thLetterSpacing: PropTypes.string,
-    columns: PropTypes.array
+    thBoxShadow: PropTypes.string,
+    columns: PropTypes.array,
+    newDesign: PropTypes.bool
   };
+
+  get style() {
+    return this.props.newDesign ? newStyle : oldStyle;
+  }
 
   renderSortingArrow = (sortDescending, colNum) => {
     if (sortDescending === undefined) {
       return null;
     }
-    const sortDirectionClassName = sortDescending ? s.sortArrowAsc : s.sortArrowDesc;
+    if (this.props.newDesign) {
+      const Arrow = sortDescending ? SortByArrowUp : SortByArrowDown;
+      return (
+        <span
+          data-hook={`${colNum}_title`}
+          className={this.style.sortArrow}
+          >
+          <Arrow height={12} data-hook={sortDescending ? 'sort_arrow_dec' : 'sort_arrow_asc'}/>
+        </span>);
+    }
+    const sortDirectionClassName = sortDescending ? this.style.sortArrowAsc : this.style.sortArrowDesc;
     return <span data-hook={`${colNum}_title`} className={sortDirectionClassName}><ArrowVertical/></span>;
   };
 
@@ -251,11 +281,12 @@ class TableHeader extends Component {
       moveBy: {x: 2.5, y: -7}
     });
     return (
-      <div onClick={e => e.stopPropagation()}>
-        <Tooltip {..._tooltipProps}>
-          <span><InfoCircle className={s.infoIcon} size={24}/></span>
-        </Tooltip>
-      </div>);
+      <Tooltip {..._tooltipProps}>
+        <span className={this.style.infoTooltipWrapper}>
+          <InfoCircle className={this.style.infoIcon} size={24}/>
+        </span>
+      </Tooltip>
+    );
   };
 
   renderHeaderCell = (column, colNum) => {
@@ -265,6 +296,7 @@ class TableHeader extends Component {
       height: this.props.thHeight,
       fontSize: this.props.thFontSize,
       border: this.props.thBorder,
+      boxShadow: this.props.thBoxShadow,
       color: this.props.thColor,
       opacity: this.props.thOpacity,
       letterSpacing: this.props.thLetterSpacing,
@@ -279,9 +311,10 @@ class TableHeader extends Component {
       <th
         style={style}
         key={colNum}
+        className={classNames({[typography.t4]: this.props.newDesign})}
         {...optionalHeaderCellProps}
         >
-        <div className={s.thContainer}>
+        <div className={this.style.thContainer}>
           {column.title}{this.renderSortingArrow(column.sortDescending, colNum)}{this.renderInfoTooltip(column.infoTooltip, colNum)}
         </div>
       </th>);
@@ -320,9 +353,8 @@ DataTable.defaultProps = {
   loader: <div className="loader">Loading ...</div>,
   scrollElement: null,
   useWindow: true,
-  thPadding: '5px',
-  thHeight: '36px',
-  thFontSize: '12px'
+  rowVerticalPadding: 'medium',
+  newDesign: false
 };
 
 DataTable.propTypes = {
@@ -356,16 +388,43 @@ DataTable.propTypes = {
   loader: PropTypes.node,
   useWindow: PropTypes.bool,
   scrollElement: PropTypes.object,
+  rowVerticalPadding: PropTypes.oneOf([
+    'medium',
+    'large'
+  ]),
+  /** this prop is deprecated and should not be used
+   * @deprecated
+   */
   thPadding: PropTypes.string,
+  /** this prop is deprecated and should not be used
+   * @deprecated
+   */
   thHeight: PropTypes.string,
+  /** this prop is deprecated and should not be used
+   * @deprecated
+   */
   thFontSize: PropTypes.string,
+  /** this prop is deprecated and should not be used
+   * @deprecated
+   */
   thBorder: PropTypes.string,
+  /** this prop is deprecated and should not be used
+   * @deprecated
+   */
   thColor: PropTypes.string,
+  /** this prop is deprecated and should not be used
+   * @deprecated
+   */
   thOpacity: PropTypes.string,
+  /** this prop is deprecated and should not be used
+   * @deprecated
+   */
+  thBoxShadow: PropTypes.string,
   thLetterSpacing: PropTypes.string,
   rowDetails: PropTypes.func,
   allowMultiDetailsExpansion: PropTypes.bool,
-  hideHeader: PropTypes.bool
+  hideHeader: PropTypes.bool,
+  newDesign: PropTypes.bool
 };
 
 DataTable.displayName = 'DataTable';