+++ stories/DataTable/Example.js
@@ -3,7 +3,7 @@ import DataTable from 'wix-style-react/DataTable';
 import s from './Example.scss';
 
 const style = {
-  width: '50%'
+  width: '966px'
 };
 
 const baseData = [
@@ -26,7 +26,7 @@ class DataTableExample extends React.Component {
     return (
       <div style={style}>
         <DataTable
-          dataHook="story-data-table"
+          dataHook="story-data-table-example"
           data={generateData()}
           onRowClick={(row, rowNum) => {
             /*eslint-disable no-alert*/
@@ -35,6 +35,7 @@ class DataTableExample extends React.Component {
           }}
           dynamicRowClass={row => row.firstName === baseData[1].firstName ? s.highlightRow : null}
           infiniteScroll
+          newDesign
           itemsPerPage={20}
           columns={[
               {title: 'Row Number', render: (row, rowNum) => '#' + (rowNum + 1), width: '20%', minWidth: '75px', important: true},