+++ stories/DataTable/ExampleWithAnimatedRowDetails.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import './Example.scss';
 
 const style = {
-  width: '50%'
+  width: '966px'
 };
 
 const baseData = [
@@ -46,6 +46,7 @@ class ExampleWithAnimatedRowDetails extends React.Component {
           data={generateData()}
           rowDetails={row => <MyRowDetailsComponent {...row}/>}
           rowDetailsAnimation
+          newDesign
           allowMultiDetailsExpansion
           columns={[
             {title: 'Row Number', render: (row, rowNum) => '#' + (rowNum + 1), width: '20%', minWidth: '75px', important: true},