+++ CHANGELOG.md
@@ -23,6 +23,10 @@ Types of changes:
 - `<DatePicker/>` - fix css issues with latest `node-sass` version
 - `<Tooltip/>` - support React16 [#1814](https://github.com/wix/wix-style-react/pull/1814)
 
+### Added
+
+- `<DataTable/>` add newDesign [#1817](https://github.com/wix/wix-style-react/pull/1817)
+
 ## 4.1.3 - 2018-06-12
 ### Fixed
 - `<TextLink/>` - Updated disabled property behaciour [#1798](https://github.com/wix/wix-style-react/pull/1798)